postgresql: moved function exec_sql from the modules to module_utils (#57603)
